Separate `Sources` into a module
