+++ src/ZkFold/Symbolic/Compiler/ArithmeticCircuit/MonadBlueprint.hs
@@ -10,7 +10,7 @@ module ZkFold.Symbolic.Compiler.ArithmeticCircuit.MonadBlueprint (
     Witness,
     WitnessField,
     circuit,
-    circuits
+    circuits,
 ) where
 
 import           Control.Monad.Identity                              (Identity (..))
@@ -24,6 +24,7 @@ import           Prelude                                             hiding (Boo
                                                                       (-))
 
 import           ZkFold.Base.Algebra.Basic.Class
+import           ZkFold.Base.Algebra.Basic.Sources
 import           ZkFold.Base.Algebra.Polynomials.Multivariate        (var)
 import           ZkFold.Prelude                                      (replicate)
 import qualified ZkFold.Symbolic.Compiler.ArithmeticCircuit.Internal as I
@@ -143,46 +144,6 @@ circuits b = let (os, r) = runState b mempty in (\o -> r { acOutput = o }) <$> o
 
 sources :: forall a i . (FiniteField a, Ord i) => Witness i a -> Set i
 sources = runSources . ($ Sources @a . Set.singleton)
-
-newtype Sources a i = Sources { runSources :: Set i } deriving newtype (Semigroup, Monoid)
-
-instance MultiplicativeSemigroup c => Exponent (Sources a i) c where
-  (^) = const
-
-instance MultiplicativeMonoid c => Scale c (Sources a i) where
-  scale = const id
-
-instance Ord i => AdditiveSemigroup (Sources a i) where
-  (+) = (<>)
-
-instance Ord i => AdditiveMonoid (Sources a i) where
-  zero = mempty
-
-instance Ord i => AdditiveGroup (Sources a i) where
-  negate = id
-
-instance Finite a => Finite (Sources a i) where
-  type Order (Sources a i) = Order a
-
-instance Ord i => MultiplicativeSemigroup (Sources a i) where
-  (*) = (<>)
-
-instance Ord i => MultiplicativeMonoid (Sources a i) where
-  one = mempty
-
-instance Ord i => MultiplicativeGroup (Sources a i) where
-  invert = id
-
-instance Ord i => FromConstant c (Sources a i) where
-  fromConstant _ = mempty
-
-instance Ord i => Semiring (Sources a i)
-
-instance Ord i => Ring (Sources a i)
-
-instance Ord i => Field (Sources a i) where
-    finv = id
-    rootOfUnity _ = Just (Sources mempty)
 
 instance (Finite a, Ord i) => BinaryExpansion (Sources a i) where
   binaryExpansion = replicate (numberOfBits @a)
