+++ zkfold-base.cabal
@@ -81,6 +81,7 @@ library
     import: options
     exposed-modules:
       ZkFold.Base.Algebra.Basic.Class
+      ZkFold.Base.Algebra.Basic.Sources
       ZkFold.Base.Algebra.Basic.Field
       ZkFold.Base.Algebra.Basic.Number
       ZkFold.Base.Algebra.Basic.Permutations
